TIP-810: Remove registryInterface and move registry
