--- src/Akeneo/Bundle/ElasticsearchBundle/Registry/ClientRegistryInterface.php
@@ -1,23 +1,0 @@
-<?php
-
-namespace Akeneo\Bundle\ElasticsearchBundle\Registry;
-
-use Akeneo\Bundle\ElasticsearchBundle\Client;
-
-/**
- * Registry of Elasticsearch clients.
- */
-interface ClientRegistryInterface
-{
-    /**
-     * @param Client $client
-     *
-     * @return mixed
-     */
-    public function register(Client $client): void;
-
-    /**
-     * @return Client[]
-     */
-    public function getClients(): array;
-}
