+++ src/Akeneo/Bundle/ElasticsearchBundle/Resources/config/services.yml
@@ -1,6 +1,6 @@
 parameters:
     akeneo_elasticsearch.client_builder.class: 'Elasticsearch\ClientBuilder'
-    akeneo_elasticsearch.registry.clients.class: Akeneo\Bundle\ElasticsearchBundle\Registry\ClientRegistry
+    akeneo_elasticsearch.registry.clients.class: Akeneo\Bundle\ElasticsearchBundle\ClientRegistry
 
 services:
     akeneo_elasticsearch.registry.clients:
